Preview the claim-impound states from one class
Inherits Wrapper::ComponentPreview for #page rather than sharing it through a
module, so the nine preview files collapse to one.
