+++ app/components/registrations/show/wrapper/component_preview.rb
@@ -6,10 +6,9 @@ module Registrations
       # The whole registration show page, one scenario per overlay it can raise. These
       # render a persisted bike — the page is far too query-heavy for an in-memory one — so
       # they're gated out of production, where the bikes would be someone's real ones.
-      # The claim-impound card's states are their own previews, under ClaimImpound::
+      # The claim-impound card's states preview from ClaimImpound::ComponentPreview,
+      # which inherits #page from here
       class ComponentPreview < ApplicationComponentPreview
-        include PreviewPage
-
         # @param view select [consumer, org_admin]
         # @param bike_id text "Bike to render — defaults to one of the org's"
         def no_overlay(view: "consumer", bike_id: nil)
@@ -77,6 +76,50 @@ module Registrations
         end
 
         private
+
+        # The block builds the alerts off the resolved bike, or returns :missing.
+        # as_view pins the perspective for a scenario that only makes sense in one
+        def page(view:, bike_id:, bike_sticker: nil, current_user: lookbook_user, as_view: nil)
+          return production_notice("registration") if Rails.env.production?
+
+          bike = preview_bike(bike_id)
+          return missing_notice("a bike") if bike.blank?
+
+          current_alerts = block_given? ? yield(bike) : {}
+          return missing_notice("the record this scenario needs") if current_alerts == :missing
+
+          available_views = ::BikeServices::ShowViews.available(bike:, current_user:,
+            organization: lookbook_organization)
+          resolved = as_view || resolved_view(view, available_views, bike:, current_user:)
+          component = Component.new(bike:, current_user:, view: resolved, available_views:,
+            bike_sticker:, current_alerts:)
+
+          render_with_template(template: "registrations/show/wrapper/preview/scenario",
+            locals: {component:, offset_header: resolved.last.nil?})
+        end
+
+        # ShowViews decides what this viewer may see; the param only picks which side of
+        # that to show, so a preview can't put up a page the app never serves
+        def resolved_view(view, available_views, bike:, current_user:)
+          org_view = view.to_s == "org_admin"
+          available_views.find { |_kind, organization| organization.present? == org_view } ||
+            ::BikeServices::ShowViews.default_view_for(bike:, current_user:, organization: lookbook_organization)
+        end
+
+        # Claimed by default — SentToNewOwner raises itself off an unclaimed registration,
+        # stacking that alert onto every other scenario
+        def preview_bike(bike_id)
+          ::Bike.unscoped.find_by(id: bike_id) || bike_with_ownership(claimed: true)
+        end
+
+        def bike_with_ownership(claimed:)
+          owned = ::Ownership.current.where(claimed:).select(:bike_id)
+          org_bikes.where(id: owned).last || ::Bike.unscoped.where(id: owned).last
+        end
+
+        def org_bikes
+          lookbook_organization&.bikes || ::Bike.none
+        end
 
         def parking_notification(bike)
           sent = ::ParkingNotification.where.not(retrieval_link_token: nil)
